added hooks for specific web app to add its own processing for better code structure
