+++ hs_gabbs_web_app/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'hs_gabbs_web_app.apps.HSGABBsAppConfig'
