+++ hs_tracking/signals.py
@@ -1,11 +1,14 @@
 from django.contrib.auth.signals import user_logged_in, user_logged_out
-from django.dispatch import receiver
+from django.dispatch import receiver, Signal
 
 from hs_core.signals import pre_download_file, post_delete_resource, post_create_resource
 
 from .models import Session
 from .models import Variable
 from .utils import get_std_log_fields
+
+
+post_tracking_web_app_launch_url = Signal(providing_args=['request', 'url'])
 
 
 @receiver(user_logged_in, dispatch_uid='id_capture_login')
